simd: squashed updates

+++ kernels/cuda/moving_averages/otto_kernel.cu
@@ -40,66 +40,66 @@ void otto_batch_f32(
     const int p2 = slow;
     const int p3 = max(slow * fast, 1);
 
-    const double a1_base = 2.0 / (static_cast<double>(p1) + 1.0);
-    const double a2_base = 2.0 / (static_cast<double>(p2) + 1.0);
-    const double a3_base = 2.0 / (static_cast<double>(p3) + 1.0);
+    const float a1_base = 2.0f / (static_cast<float>(p1) + 1.0f);
+    const float a2_base = 2.0f / (static_cast<float>(p2) + 1.0f);
+    const float a3_base = 2.0f / (static_cast<float>(p3) + 1.0f);
 
     const int ott_p = max(__ldg(ott_periods + combo), 1);
-    const double a_base_lott = 2.0 / (static_cast<double>(ott_p) + 1.0);
-    const double ott_percent = static_cast<double>(__ldg(ott_percents + combo));
-    const double coco = static_cast<double>(__ldg(cocos + combo));
+    const float a_base_lott = 2.0f / (static_cast<float>(ott_p) + 1.0f);
+    const float ott_percent = __ldg(ott_percents + combo);
+    const float coco = __ldg(cocos + combo);
 
-    const double fark = ott_percent * 0.01;
-    const double scale_up = (200.0 + ott_percent) / 200.0;
-    const double scale_dn = (200.0 - ott_percent) / 200.0;
+    const float fark = ott_percent * 0.01f;
+    const float scale_up = (200.0f + ott_percent) / 200.0f;
+    const float scale_dn = (200.0f - ott_percent) / 200.0f;
 
     float* __restrict__ hott_row = hott_out + combo * series_len;
     float* __restrict__ lott_row = lott_out + combo * series_len;
 
     // VIDYA tracks for source
-    double v1 = 0.0, v2 = 0.0, v3 = 0.0;
+    float v1 = 0.0f, v2 = 0.0f, v3 = 0.0f;
     // For VAR(LOTT): CMO(9) on lott diffs
     const int CMO_P = 9;
-    double ring_up[CMO_P];
-    double ring_dn[CMO_P];
+    float ring_up[CMO_P];
+    float ring_dn[CMO_P];
     #pragma unroll
-    for (int k = 0; k < CMO_P; ++k) { ring_up[k] = 0.0; ring_dn[k] = 0.0; }
-    double sum_up = 0.0, sum_dn = 0.0;
+    for (int k = 0; k < CMO_P; ++k) { ring_up[k] = 0.0f; ring_dn[k] = 0.0f; }
+    float sum_up = 0.0f, sum_dn = 0.0f;
     int head = 0;
 
-    double prev_lott = 0.0;
-    double ma_prev = 0.0;
-    double long_stop_prev = NAN, short_stop_prev = NAN;
+    float prev_lott = 0.0f;
+    float ma_prev = 0.0f;
+    float long_stop_prev = NAN, short_stop_prev = NAN;
     int dir_prev = 1;
 
     for (int i = 0; i < series_len; ++i) {
-        const double x = nzf(__ldg(prices + i));
-        const double c_abs = static_cast<double>(__ldg(cabs + i));
+        const float x = nzf(__ldg(prices + i));
+        const float c_abs = __ldg(cabs + i);
 
         // Adaptive alphas
-        const double a1 = a1_base * c_abs;
-        const double a2 = a2_base * c_abs;
-        const double a3 = a3_base * c_abs;
+        const float a1 = a1_base * c_abs;
+        const float a2 = a2_base * c_abs;
+        const float a3 = a3_base * c_abs;
 
         // VIDYA on source
-        v1 = fma(a1, x, (1.0 - a1) * v1);
-        v2 = fma(a2, x, (1.0 - a2) * v2);
-        v3 = fma(a3, x, (1.0 - a3) * v3);
+        v1 = fmaf(a1, x, (1.0f - a1) * v1);
+        v2 = fmaf(a2, x, (1.0f - a2) * v2);
+        v3 = fmaf(a3, x, (1.0f - a3) * v3);
 
         // LOTT = v1 / ((v2 - v3) + coco)
-        const double denom_l = (v2 - v3) + coco;
-        const double lott = denom_l != 0.0 ? (v1 / denom_l) : 0.0;
-        lott_row[i] = static_cast<float>(lott);
+        const float denom_l = (v2 - v3) + coco;
+        const float lott = denom_l != 0.0f ? (v1 / denom_l) : 0.0f;
+        lott_row[i] = lott;
 
         // CMO(9) on LOTT diffs (VIDYA on LOTT)
         if (i > 0) {
-            const double d = lott - prev_lott;
+            const float d = lott - prev_lott;
             if (i >= CMO_P) {
                 sum_up -= ring_up[head];
                 sum_dn -= ring_dn[head];
             }
-            const double up = d > 0.0 ? d : 0.0;
-            const double dn = d > 0.0 ? 0.0 : -d;
+            const float up = d > 0.0f ? d : 0.0f;
+            const float dn = d > 0.0f ? 0.0f : -d;
             ring_up[head] = up;
             ring_dn[head] = dn;
             sum_up += up;
@@ -108,27 +108,27 @@ void otto_batch_f32(
         }
         prev_lott = lott;
 
-        const double denom = sum_up + sum_dn;
-        const double c2 = (i >= CMO_P && denom != 0.0) ? fabs((sum_up - sum_dn) / denom) : 0.0;
-        const double a_lott = a_base_lott * c2;
-        const double ma = fma(a_lott, lott, (1.0 - a_lott) * ma_prev);
+        const float denom = sum_up + sum_dn;
+        const float c2 = (i >= CMO_P && denom != 0.0f) ? fabsf((sum_up - sum_dn) / denom) : 0.0f;
+        const float a_lott = a_base_lott * c2;
+        const float ma = fmaf(a_lott, lott, (1.0f - a_lott) * ma_prev);
         ma_prev = ma;
 
         if (i == 0) {
-            long_stop_prev = ma * (1.0 - fark);
-            short_stop_prev = ma * (1.0 + fark);
-            const double mt = long_stop_prev;
-            hott_row[i] = static_cast<float>(ma > mt ? mt * scale_up : mt * scale_dn);
+            long_stop_prev = ma * (1.0f - fark);
+            short_stop_prev = ma * (1.0f + fark);
+            const float mt = long_stop_prev;
+            hott_row[i] = (ma > mt ? mt * scale_up : mt * scale_dn);
         } else {
-            const double ls = ma * (1.0 - fark);
-            const double ss = ma * (1.0 + fark);
-            const double long_stop = (ma > long_stop_prev) ? fmax(ls, long_stop_prev) : ls;
-            const double short_stop = (ma < short_stop_prev) ? fmin(ss, short_stop_prev) : ss;
+            const float ls = ma * (1.0f - fark);
+            const float ss = ma * (1.0f + fark);
+            const float long_stop = (ma > long_stop_prev) ? fmaxf(ls, long_stop_prev) : ls;
+            const float short_stop = (ma < short_stop_prev) ? fminf(ss, short_stop_prev) : ss;
             const int dir = (dir_prev == -1 && ma > short_stop_prev)
                                 ? 1
                                 : ((dir_prev == 1 && ma < long_stop_prev) ? -1 : dir_prev);
-            const double mt = (dir == 1) ? long_stop : short_stop;
-            hott_row[i] = static_cast<float>(ma > mt ? mt * scale_up : mt * scale_dn);
+            const float mt = (dir == 1) ? long_stop : short_stop;
+            hott_row[i] = (ma > mt ? mt * scale_up : mt * scale_dn);
             long_stop_prev = long_stop;
             short_stop_prev = short_stop;
             dir_prev = dir;
@@ -155,93 +155,92 @@ void otto_many_series_one_param_f32(
     const int p1 = max(slow_vidya / 2, 1);
     const int p2 = max(slow_vidya, 1);
     const int p3 = max(slow_vidya * max(fast_vidya, 1), 1);
-    const double a1_base = 2.0 / (static_cast<double>(p1) + 1.0);
-    const double a2_base = 2.0 / (static_cast<double>(p2) + 1.0);
-    const double a3_base = 2.0 / (static_cast<double>(p3) + 1.0);
-    const double a_base_lott = 2.0 / (static_cast<double>(max(ott_period, 1)) + 1.0);
-    const double coco = static_cast<double>(coco_f);
-    const double ott_percent = static_cast<double>(ott_percent_f);
-    const double fark = ott_percent * 0.01;
-    const double scale_up = (200.0 + ott_percent) / 200.0;
-    const double scale_dn = (200.0 - ott_percent) / 200.0;
+    const float a1_base = 2.0f / (static_cast<float>(p1) + 1.0f);
+    const float a2_base = 2.0f / (static_cast<float>(p2) + 1.0f);
+    const float a3_base = 2.0f / (static_cast<float>(p3) + 1.0f);
+    const float a_base_lott = 2.0f / (static_cast<float>(max(ott_period, 1)) + 1.0f);
+    const float coco = coco_f;
+    const float ott_percent = ott_percent_f;
+    const float fark = ott_percent * 0.01f;
+    const float scale_up = (200.0f + ott_percent) / 200.0f;
+    const float scale_dn = (200.0f - ott_percent) / 200.0f;
 
     // Precompute c_abs on price locally (CMO 9)
     const int CMO_P = 9;
-    double ring_up_p[CMO_P];
-    double ring_dn_p[CMO_P];
+    float ring_up_p[CMO_P];
+    float ring_dn_p[CMO_P];
     #pragma unroll
-    for (int k = 0; k < CMO_P; ++k) { ring_up_p[k] = 0.0; ring_dn_p[k] = 0.0; }
-    double sum_up_p = 0.0, sum_dn_p = 0.0; int head_p = 0;
+    for (int k = 0; k < CMO_P; ++k) { ring_up_p[k] = 0.0f; ring_dn_p[k] = 0.0f; }
+    float sum_up_p = 0.0f, sum_dn_p = 0.0f; int head_p = 0;
 
-    double v1 = 0.0, v2 = 0.0, v3 = 0.0;
-    double prev_price = 0.0;
+    float v1 = 0.0f, v2 = 0.0f, v3 = 0.0f;
+    float prev_price = 0.0f;
 
     // For VAR(LOTT)
-    double ring_up_l[CMO_P];
-    double ring_dn_l[CMO_P];
+    float ring_up_l[CMO_P];
+    float ring_dn_l[CMO_P];
     #pragma unroll
-    for (int k = 0; k < CMO_P; ++k) { ring_up_l[k] = 0.0; ring_dn_l[k] = 0.0; }
-    double sum_up_l = 0.0, sum_dn_l = 0.0; int head_l = 0;
-    double prev_lott = 0.0;
-    double ma_prev = 0.0;
-    double long_stop_prev = NAN, short_stop_prev = NAN; int dir_prev = 1;
+    for (int k = 0; k < CMO_P; ++k) { ring_up_l[k] = 0.0f; ring_dn_l[k] = 0.0f; }
+    float sum_up_l = 0.0f, sum_dn_l = 0.0f; int head_l = 0;
+    float prev_lott = 0.0f;
+    float ma_prev = 0.0f;
+    float long_stop_prev = NAN, short_stop_prev = NAN; int dir_prev = 1;
 
     for (int t = 0; t < cols; ++t) {
-        const double x = nzf(prices_tm[t * rows + series]);
+        const float x = nzf(prices_tm[t * rows + series]);
         if (t > 0) {
-            const double d = x - prev_price;
+            const float d = x - prev_price;
             if (t >= CMO_P) { sum_up_p -= ring_up_p[head_p]; sum_dn_p -= ring_dn_p[head_p]; }
-            const double up = d > 0.0 ? d : 0.0;
-            const double dn = d > 0.0 ? 0.0 : -d;
+            const float up = d > 0.0f ? d : 0.0f;
+            const float dn = d > 0.0f ? 0.0f : -d;
             ring_up_p[head_p] = up; ring_dn_p[head_p] = dn;
             sum_up_p += up; sum_dn_p += dn; head_p = (head_p + 1) == CMO_P ? 0 : (head_p + 1);
         }
         prev_price = x;
-        const double denom_p = sum_up_p + sum_dn_p;
-        const double c_abs = (t >= CMO_P && denom_p != 0.0) ? fabs((sum_up_p - sum_dn_p)/denom_p) : 0.0;
+        const float denom_p = sum_up_p + sum_dn_p;
+        const float c_abs = (t >= CMO_P && denom_p != 0.0f) ? fabsf((sum_up_p - sum_dn_p) / denom_p) : 0.0f;
 
-        const double a1 = a1_base * c_abs;
-        const double a2 = a2_base * c_abs;
-        const double a3 = a3_base * c_abs;
-        v1 = fma(a1, x, (1.0 - a1) * v1);
-        v2 = fma(a2, x, (1.0 - a2) * v2);
-        v3 = fma(a3, x, (1.0 - a3) * v3);
-        const double denom_l = (v2 - v3) + coco;
-        const double lott = denom_l != 0.0 ? (v1 / denom_l) : 0.0;
-        lott_tm[t * rows + series] = static_cast<float>(lott);
+        const float a1 = a1_base * c_abs;
+        const float a2 = a2_base * c_abs;
+        const float a3 = a3_base * c_abs;
+        v1 = fmaf(a1, x, (1.0f - a1) * v1);
+        v2 = fmaf(a2, x, (1.0f - a2) * v2);
+        v3 = fmaf(a3, x, (1.0f - a3) * v3);
+        const float denom_l = (v2 - v3) + coco;
+        const float lott = denom_l != 0.0f ? (v1 / denom_l) : 0.0f;
+        lott_tm[t * rows + series] = lott;
 
         if (t > 0) {
-            const double d = lott - prev_lott;
+            const float d = lott - prev_lott;
             if (t >= CMO_P) { sum_up_l -= ring_up_l[head_l]; sum_dn_l -= ring_dn_l[head_l]; }
-            const double up = d > 0.0 ? d : 0.0;
-            const double dn = d > 0.0 ? 0.0 : -d;
+            const float up = d > 0.0f ? d : 0.0f;
+            const float dn = d > 0.0f ? 0.0f : -d;
             ring_up_l[head_l] = up; ring_dn_l[head_l] = dn;
             sum_up_l += up; sum_dn_l += dn; head_l = (head_l + 1) == CMO_P ? 0 : (head_l + 1);
         }
         prev_lott = lott;
-        const double denom_lc = sum_up_l + sum_dn_l;
-        const double c2 = (t >= CMO_P && denom_lc != 0.0) ? fabs((sum_up_l - sum_dn_l)/denom_lc) : 0.0;
-        const double a_lott = a_base_lott * c2;
-        const double ma = fma(a_lott, lott, (1.0 - a_lott) * ma_prev);
+        const float denom_lc = sum_up_l + sum_dn_l;
+        const float c2 = (t >= CMO_P && denom_lc != 0.0f) ? fabsf((sum_up_l - sum_dn_l) / denom_lc) : 0.0f;
+        const float a_lott = a_base_lott * c2;
+        const float ma = fmaf(a_lott, lott, (1.0f - a_lott) * ma_prev);
         ma_prev = ma;
 
         if (t == 0) {
-            long_stop_prev = ma * (1.0 - fark);
-            short_stop_prev = ma * (1.0 + fark);
-            const double mt = long_stop_prev;
-            hott_tm[t * rows + series] = static_cast<float>(ma > mt ? mt * scale_up : mt * scale_dn);
+            long_stop_prev = ma * (1.0f - fark);
+            short_stop_prev = ma * (1.0f + fark);
+            const float mt = long_stop_prev;
+            hott_tm[t * rows + series] = (ma > mt ? mt * scale_up : mt * scale_dn);
         } else {
-            const double ls = ma * (1.0 - fark);
-            const double ss = ma * (1.0 + fark);
-            const double long_stop = (ma > long_stop_prev) ? fmax(ls, long_stop_prev) : ls;
-            const double short_stop = (ma < short_stop_prev) ? fmin(ss, short_stop_prev) : ss;
+            const float ls = ma * (1.0f - fark);
+            const float ss = ma * (1.0f + fark);
+            const float long_stop = (ma > long_stop_prev) ? fmaxf(ls, long_stop_prev) : ls;
+            const float short_stop = (ma < short_stop_prev) ? fminf(ss, short_stop_prev) : ss;
             const int dir = (dir_prev == -1 && ma > short_stop_prev)
                                 ? 1
                                 : ((dir_prev == 1 && ma < long_stop_prev) ? -1 : dir_prev);
-            const double mt = (dir == 1) ? long_stop : short_stop;
-            hott_tm[t * rows + series] = static_cast<float>(ma > mt ? mt * scale_up : mt * scale_dn);
+            const float mt = (dir == 1) ? long_stop : short_stop;
+            hott_tm[t * rows + series] = (ma > mt ? mt * scale_up : mt * scale_dn);
             long_stop_prev = long_stop; short_stop_prev = short_stop; dir_prev = dir;
         }
     }
 }
-
